Biologics UI Components - AliasInput, renderers and schemas.ts
- Factor AliasInput into its own file
- export resolveRenderer, CBMB, ASSAY_TABLES, EXP_TABLES, DATA_CLASSES, SAMPLE_SETS
